feat: initialise huma API
Adds a huma/echo API server. Simply has a probe endpoint right now
for checking that the server is started up.

Adds a base framework around http controllers that can be used to
add more quite easily later.

+++ pkg/config/main.go
@@ -6,7 +6,7 @@ type ServerAccessLogsConfig struct {
 }
 
 type ServerConfig struct {
-	Port int
+	Port uint16
 	AccessLogs ServerAccessLogsConfig `yaml:"accessLogs" mapstructure:"access_logs"`
 }
 
@@ -20,7 +20,7 @@ type Config struct {
 	Logging LoggingConfig
 }
 
-func (c *Config) GetServerPort() int {
+func (c *Config) GetServerPort() uint16 {
 	return c.Server.Port
 }
